pandas/requirements-dev.txt
# This file is auto-generated from environment.yml, do not modify.
# See that file for comments about the need/usage of each dependency.

pip
versioneer[toml]
cython~=3.0.5
meson[ninja]==1.2.1
meson-python==0.13.1
pytest>=7.3.2
pytest-cov
pytest-xdist>=3.4.0
pytest-qt>=4.4.0
pytest-localserver
PyQt5>=5.15.9
coverage
python-dateutil
numpy<2
beautifulsoup4>=4.11.2
blosc
bottleneck>=1.3.6
fastparquet>=2023.10.0
fsspec>=2022.11.0
html5lib>=1.1
hypothesis>=6.84.0
gcsfs>=2022.11.0
ipython
pickleshare
jinja2>=3.1.2
lxml>=4.9.2
matplotlib>=3.6.3
numba>=0.56.4
numexpr>=2.8.4
openpyxl>=3.1.0
odfpy>=1.4.1
py
psycopg2-binary>=2.9.6
pyarrow>=10.0.1
pymysql>=1.0.2
pyreadstat>=1.2.0
tables>=3.8.0
python-calamine>=0.1.7
pytz>=2023.4
pyxlsb>=1.0.10
s3fs>=2022.11.0
scipy>=1.10.0
SQLAlchemy>=2.0.0
tabulate>=0.9.0
xarray>=2022.12.0, <=2024.9.0
xlrd>=2.0.1
xlsxwriter>=3.0.5
zstandard>=0.19.0
dask
seaborn
moto
flask
asv>=0.6.1
flake8==7.1.0
mypy==1.13.0
tokenize-rt
pre-commit>=4.0.1
gitpython
gitdb
google-auth
natsort
numpydoc
pydata-sphinx-theme==0.16
pytest-cython
sphinx
sphinx-design
sphinx-copybutton
types-python-dateutil
types-PyMySQL
types-pytz
types-PyYAML
types-setuptools
nbconvert>=7.11.0
nbsphinx
pandoc
ipywidgets
nbformat
notebook>=7.0.6
ipykernel
markdown
feedparser
pyyaml
requests
pygments
adbc-driver-postgresql>=0.10.0
adbc-driver-sqlite>=0.8.0
typing_extensions; python_version<"3.11"
tzdata>=2022.7
